pyproject.toml
[project]
name = "projeto-carteira-digital"
version = "0.1.0"
description = "API de Carteira Digital da disciplina Projeto Banco de Dados:"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.121.3",
    "httpx>=0.28.1",
    "mysql>=0.0.3",
    "mysql-connector-python>=9.5.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]